refactor: extract shared demo commands to eliminate duplication
Move EchoCommand and UpperCommand into DemoCommands utility class
shared by ShellJobExample and ShellPipelineExample, removing the
duplicated inner class definitions that SonarCloud flagged.
