+++ demo/src/main/java/org/jline/demo/examples/ShellPipelineExample.java
@@ -35,70 +35,6 @@ import org.jline.shell.impl.SimpleCommandGroup;
 public class ShellPipelineExample {
 
     // SNIPPET_START: ShellPipelineExample
-    static class EchoCommand extends AbstractCommand {
-        EchoCommand() {
-            super("echo");
-        }
-
-        @Override
-        public String description() {
-            return "Echo arguments to output";
-        }
-
-        /**
-         * Prints the command arguments joined by spaces to the session output.
-         *
-         * @param session the command session used for input/output
-         * @param args the arguments to join and print
-         * @return the joined arguments string
-         */
-        @Override
-        public Object execute(CommandSession session, String[] args) {
-            String msg = String.join(" ", args);
-            session.out().println(msg);
-            return msg;
-        }
-    }
-
-    static class UpperCommand extends AbstractCommand {
-        UpperCommand() {
-            super("upper");
-        }
-
-        /**
-         * Provide the one-line help text for the command.
-         *
-         * @return the one-line help text describing the command's behavior: "Convert pipe input to upper case"
-         */
-        @Override
-        public String description() {
-            return "Convert pipe input to upper case";
-        }
-
-        /**
-         * Converts input text to uppercase, writes it to the session output, and returns it.
-         *
-         * If command arguments are provided they are joined with spaces and used as the input;
-         * otherwise all available bytes from the session's input stream are read and trimmed.
-         *
-         * @param session the command session providing input and output streams
-         * @param args    command arguments; when non-empty they are joined into the input text
-         * @return        the resulting uppercase string
-         */
-        @Override
-        public Object execute(CommandSession session, String[] args) throws Exception {
-            String input;
-            if (args.length > 0) {
-                input = String.join(" ", args);
-            } else {
-                input = new String(session.in().readAllBytes()).trim();
-            }
-            String result = input.toUpperCase();
-            session.out().println(result);
-            return result;
-        }
-    }
-
     static class FailCommand extends AbstractCommand {
         FailCommand() {
             super("fail");
@@ -123,7 +59,8 @@ public class ShellPipelineExample {
                 .prompt("pipeline-demo> ")
                 .pipelineParser(customParser) // HIGHLIGHT
                 .helpCommands(true)
-                .groups(new SimpleCommandGroup("demo", new EchoCommand(), new UpperCommand(), new FailCommand()))
+                .groups(new SimpleCommandGroup(
+                        "demo", new DemoCommands.EchoCommand(), new DemoCommands.UpperCommand(), new FailCommand()))
                 .build()) {
             shell.run();
         }
